chore: moves types to an interfaces file
